+++ src/discrete/Domain.kt
@@ -6,6 +6,7 @@ typealias Value = Int
 class Variables {
 	companion object {
 		fun <T> from(items: List<T>): List<Variable> = (0 until items.size).toList()
+		fun range(first: Int, last: Int): List<Variable> = (first..last).toList()
 	}
 }
 
@@ -13,5 +14,6 @@ class Domain {
 	companion object {
 		fun boolean(): List<Value> = listOf(1, 0)
 		fun range(first: Int, last: Int): List<Value> = (first..last).toList()
+		fun index(numItems: Int): List<Value> = (0 until numItems).toList()
 	}
 }