+++ src/fn/Common.kt
@@ -10,4 +10,13 @@ fun <T> List<Int>.mapList(values: List<T>): List<T> {
 
 fun <T> List<T>.wrapBraces(): String = "{ " + this.joinToString(separator = ", ") + " }"
 
+fun List<String>.mirroredSequence(): String {
+	if (isEmpty()) return ""
+	return if (first() > last()) {
+		reversed().joinToString(" ")
+	} else {
+		joinToString(" ")
+	}
+}
+
 fun sortedPartitionGroups(groups: List<List<String>>): List<String> = groups.map { it.sorted().wrapBraces() }