+++ src/fn/Core.kt
@@ -1,12 +1,25 @@
 package fn
 
+import discrete.Problem
 import discrete.Solution
 import discrete.SolutionCoreFn
 import discrete.Value
 
-fun <T> coreSortedPartition(values: List<Value>, items: List<T>): SolutionCoreFn {
-	return fun(solution: Solution): String {
-		val groups = solution.partitionStrings(values, items).filter { it.isNotEmpty() }
-		return sortedPartitionGroups(groups).sorted().joinToString("/")
+class CoreFn {
+	companion object {
+		fun <T> sortedPartition(values: List<Value>, items: List<T>): SolutionCoreFn {
+			return fun(solution: Solution): String {
+				val groups = solution.partitionStrings(values, items).filter { it.isNotEmpty() }
+				return sortedPartitionGroups(groups).sorted().joinToString("/")
+			}
+		}
+
+		fun <T> mirroredSequence(items: List<T>): SolutionCoreFn {
+			return fun(solution: Solution): String = solution.sequenceStrings(items).mirroredSequence()
+		}
+
+		fun <T> mirroredValues(p: Problem, items: List<T>?): SolutionCoreFn {
+			return fun(solution: Solution): String = solution.valueStrings(p, items).mirroredSequence()
+		}
 	}
 }