+++ src/fn/Objective.kt
@@ -5,13 +5,17 @@ import discrete.Score
 import discrete.Solution
 import discrete.Variable
 
-fun scoreCountUniqueValues(solution: Solution): Score = solution.values.toSet().size.toDouble()
+class ScoreFn {
+	companion object {
+		fun countUniqueValues(solution: Solution): Score = solution.values.toSet().size.toDouble()
 
-fun scoreSubsetSize(solution: Solution): Score = solution.values.sum().toDouble()
+		fun subsetSize(solution: Solution): Score = solution.values.sum().toDouble()
 
-fun scoreSumWeightedValues(variables: List<Variable>, weight: List<Double>): ObjectiveFn {
-	return fun(solution: Solution): Score {
-		val count = solution.map
-		return variables.sumOf { (count[it] ?: 0) * weight[it] }
+		fun sumWeightedValues(variables: List<Variable>, weight: List<Double>): ObjectiveFn {
+			return fun(solution: Solution): Score {
+				val count = solution.map
+				return variables.sumOf { (count[it] ?: 0) * weight[it] }
+			}
+		}
 	}
 }